pyproject.toml
[tool.poetry]
name = "json-server.py"
version = "0.1.5"
description = "A simple JSON server"
authors = ["Gerald <gera2ld@163.com>"]
license = "MIT"
packages = [
  { include = "json_server" },
]
repository = 'https://github.com/gera2ld/json-server.py'
readme = 'README.md'

[tool.poetry.dependencies]
python = "^3.6"
click = "^7.0"
aiohttp = "^3.6.2"
gera2ld-pyserve = "^0.1.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry.scripts]
json-server = "json_server.cli:main"
